Split triggers in e2e emulator integration test into many codebases.
