+++ scripts/triggers-end-to-end-tests/v2/.gitignore
@@ -1,0 +1,3 @@
+node_modules/
+.eslintrc
+package-lock.json
